% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Useful_Functions_Definitions.R
\name{herb_strsplit}
\alias{herb_strsplit}
\title{Split a string and un-list the vector
Split a string}
\usage{
herb_strsplit(x, split)
}
\arguments{
\item{x}{A character vector with one element.}

\item{split}{What to split on.}
}
\value{
A character vector.
}
\description{
Split a string and un-list the vector
Split a string
}
\examples{
x <- "alfa,bravo,charlie,delta"
herb_strsplit(x, split = ",")

}
